refactor(mini-bar): extract shared SessionDropdown component
