+++ src/components/StatusPill.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useLayoutEffect, useRef, useState } from "react";
 import { createPortal } from "react-dom";
-import { invoke } from "@tauri-apps/api/core";
 import { listen } from "@tauri-apps/api/event";
 import {
   getCurrentWindow,
@@ -15,11 +14,10 @@ import {
   detectHome,
   reflectActiveServices,
   overlayClaudeState,
-  groupBasename,
-  shellLabel,
   type Group,
 } from "../utils/sessionGroups";
 import { useSessionList } from "../hooks/useSessionList";
+import { SessionDropdown } from "./SessionDropdown";
 import { useSessionGroupCount } from "../hooks/useSessionGroupCount";
 import { useLanList } from "../hooks/useLanList";
 import { useOpacity } from "../hooks/useOpacity";
@@ -438,107 +436,15 @@ export function StatusPill({ status, glow, disabled = false, onOpenChange, onMin
       </div>
 
       {sessionListEnabled && sessionOpen && (
-        <div
-          data-testid="session-dropdown"
-          className="session-dropdown"
-          role="menu"
-          style={{
-            top: `${dropdownTop}px`,
-            maxHeight: `${dropdownMaxHeight}px`,
-          }}
-        >
-          {groups.length === 0 ? (
-            <div className="session-empty">No active terminals</div>
-          ) : (
-            groups.map((g) => {
-              const isCollapsed = collapsed.has(g.key);
-              const headContent = (
-                <>
-                  {!g.isClaudeFallback && (
-                    <span className="session-group-caret" aria-hidden="true" />
-                  )}
-                  <span className={`dot small ${g.state}`} />
-                  <span className="session-group-title-row">
-                    <span className="session-group-title">
-                      {groupBasename(g)}
-                    </span>
-                    {g.pretty && g.pretty !== groupBasename(g) && (
-                      <span
-                        className="session-group-info"
-                        aria-label={`Full path: ${g.pretty}`}
-                        onMouseEnter={(e) =>
-                          showPathTooltip(e.currentTarget, g.pretty)
-                        }
-                        onMouseLeave={hidePathTooltip}
-                      >
-                        ?
-                      </span>
-                    )}
-                  </span>
-                </>
-              );
-              return (
-                <div
-                  key={g.key}
-                  className={`session-group ${g.isClaudeFallback ? "claude" : ""}`}
-                  data-testid={`session-group-${g.key}`}
-                >
-                  {g.isClaudeFallback ? (
-                    <div className="session-group-head">{headContent}</div>
-                  ) : (
-                    <button
-                      type="button"
-                      className={`session-group-head clickable ${isCollapsed ? "collapsed" : ""}`}
-                      data-testid={`session-group-head-${g.key}`}
-                      aria-expanded={!isCollapsed}
-                      aria-controls={`session-children-${g.key}`}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        void toggleCollapsed(g.key);
-                      }}
-                    >
-                      {headContent}
-                    </button>
-                  )}
-
-                  {!g.isClaudeFallback && !isCollapsed && (
-                    <div
-                      className="session-children"
-                      id={`session-children-${g.key}`}
-                    >
-                      {g.sessions.map((s) => (
-                        <button
-                          key={s.pid}
-                          type="button"
-                          className={`session-child ${s.has_claude ? "has-claude" : ""}`}
-                          data-testid={`session-item-${s.pid}`}
-                          title="Click to bring this terminal to the front"
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            invoke("focus_terminal", { pid: s.pid, tty: s.tty || null })
-                              .catch((err) => console.error("[focus_terminal]", err));
-                            setSessionOpen(false);
-                          }}
-                        >
-                          <span className={`dot small ${s.ui_state}`} />
-                          <span className="session-child-label-row">
-                            <span className="session-child-label">{shellLabel(s)}</span>
-                            {s.has_claude && (
-                              <span
-                                className="session-child-claude"
-                                aria-label="Claude Code running"
-                              />
-                            )}
-                          </span>
-                        </button>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              );
-            })
-          )}
-        </div>
+        <SessionDropdown
+          groups={groups}
+          collapsed={collapsed}
+          toggleCollapsed={(key) => void toggleCollapsed(key)}
+          onPickSession={() => setSessionOpen(false)}
+          style={{ top: `${dropdownTop}px`, maxHeight: `${dropdownMaxHeight}px` }}
+          showPathTooltip={showPathTooltip}
+          hidePathTooltip={hidePathTooltip}
+        />
       )}
 
       {pathTooltip &&
